requirements.txt
Flask==3.0.0
pytesseract==0.3.10
Pillow==10.1.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
Werkzeug==3.0.1
